+++ test/libsolidity/smtCheckerTests/control_flow/branches_with_return/simple_if_state_var.sol
@@ -1,0 +1,22 @@
+pragma experimental SMTChecker;
+
+contract C {
+
+	uint x;
+
+	function check() public {
+		require(x == 0);
+		conditional_increment();
+		assert(x == 1); // should fail;
+		assert(x == 0); // should hold;
+	}
+
+	function conditional_increment() internal {
+		if (x == 0) {
+			return;
+		}
+		x = 1;
+	}
+}
+// ----
+// Warning 6328: (132-146): CHC: Assertion violation happens here.